delete old SearchWidget and rename SearchWidgetNew to SearchWidget
