+++ lib/ui/viewer/search/search_suffix_icon_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import "package:photos/theme/ente_theme.dart";
-import "package:photos/ui/viewer/search/search_widget_new.dart";
+import 'package:photos/ui/viewer/search/search_widget.dart';
 
 class SearchSuffixIcon extends StatefulWidget {
   final bool shouldShowSpinner;
@@ -36,7 +36,7 @@ class _SearchSuffixIconState extends State<SearchSuffixIcon>
               visualDensity: const VisualDensity(horizontal: -1, vertical: -1),
               onPressed: () {
                 final searchWidgetState =
-                    context.findAncestorStateOfType<SearchWidgetNewState>()!;
+                    context.findAncestorStateOfType<SearchWidgetState>()!;
                 searchWidgetState.textController.clear();
                 searchWidgetState.focusNode.unfocus();
               },
